move folding-related classes to folding range provider
